:lipstick: コピペの色を変えた
