+++ docs/architectural.md
@@ -1,0 +1,8 @@
+# システム アーキテクチャ
+
+
+## Nuxt
+
+[参考URL](https://blog.35d.jp/2021-10-28-nuxt-composition-api-vuex-test)
+
+![composition](images/composition.png)
