+++ docs/requirements.md
@@ -182,3 +182,7 @@ prietterを.vueにも適用させる
 BEMがあまりわかっていないのかも
 メンテナンスモード
 recapthar v3を導入
+
+## nuxt コンポーネント運用について
+
+コンポーネント図で記載されている
