--- docs/uml.md
@@ -1,1 +1,0 @@
-[これ](https://qiita.com/ogomr/items/0b5c4de7f38fd1482a48)
