+++ portfolio-ui/src/components/atoms/text/VerticalText.vue
@@ -1,19 +1,32 @@
 <!-- 指定の文字を縦に配置するコンポーネント @TODO 決め打ちをやめたい -->
 <template>
   <div class="vertical">
-    <h1 class="vertical-ele">
-      <div class="">Muraflex</div>
-      <div class="">Superior</div>
-      <div class="">Demountable</div>
-      <div class="">Glass Partitions</div>
+    <h1 :class="`vertical-ele ${rotateClass}`">
+      <div v-for="(text, index) in texts" :key="index">
+        {{ text }}
+      </div>
     </h1>
   </div>
 </template>
 
 <script lang="ts">
-import { defineComponent } from '@nuxtjs/composition-api';
+import { defineComponent, PropType, ToRefs } from '@nuxtjs/composition-api';
 
 export default defineComponent({
+  props: {
+    texts: {
+      type: Object as PropType<
+        ToRefs<{
+          text: string[];
+        }>
+      >,
+      required: true
+    },
+    rotateClass: {
+      type: String,
+      default: 'transform'
+    }
+  },
   setup() {}
 });
 </script>
@@ -26,12 +39,21 @@ export default defineComponent({
   height: 100%;
 
   &-ele {
+    line-height: 1;
     white-space: nowrap;
     font-size: 4rem;
     width: 100%;
-    transform: rotate(-90deg);
     padding-left: 20%;
     margin-bottom: 0;
   }
+}
+
+/** 左に90度回転 */
+.transform {
+  transform: rotate(-90deg);
+}
+
+.transform90 {
+  transform: rotate(90deg);
 }
 </style>
